tests: rename all unit test classes, split backports
