+++ tests/unit/test_wrong_usage.py
@@ -25,11 +25,11 @@
 #
 
 '''
-test_module_setup
-~~~~~~~~~~~~~~~~~
+test_wrong_usage
+~~~~~~~~~~~~~~~~
 
 This module contains basic units tests for the setup sequence of the
-package publishing.withsphinx.
+package publishing.withsphinx to validate wrong usage.
 
 :copyright: Copyright 2014-2017 by Li-Pro.Net, see AUTHORS.
 :license: MIT, see LICENSE for details.
@@ -40,42 +40,29 @@ from __future__ import absolute_import
 from tests import util
 
 
-class TestPublishingWithSphinxSetup(util.TestCasePublishingSphinx):
+class TestWrongUsage(util.TestCasePublishingSphinx):
 
     @classmethod
     def setUpClass(self):
-        self.message = "Method should be called. Called {times} times!"
-
-    def test_package_has_setup(self):
-        '''
-        UNIT TEST: package has setup
-        '''
-        import publishing.withsphinx as pws
-        self.assertTrue(callable(pws.setup))
+        self.message = "Method should not be called. Called {times} times!"
 
     @util.mock.patch('publishing.withsphinx.backports.sphinx15')
-    def test_package_has_setup_backports(self, mock_backports_sphinx15):
-        '''
-        UNIT TEST: package has setup calling backports from Sphinx 1.5
-        '''
-        import publishing.withsphinx as pws
-        app = util.mock.MagicMock()
-        pws.setup(app)
-        self.assertTrue(mock_backports_sphinx15.called,
-                        self.message.format(times=mock_backports_sphinx15.call_count))
-        mock_backports_sphinx15.assert_called_with(app)
-
     @util.mock.patch('publishing.withsphinx.required.extensions')
-    def test_package_has_setup_required(self, mock_required_extensions):
+    def test_package_has_setup_validates_app_object(self,
+                                                    mock_backports_sphinx15,
+                                                    mock_required_extensions):
         '''
-        UNIT TEST: package has setup calling required extensions
+        UNIT TEST: package has setup validate Sphinx as application object
         '''
+        import types
         import publishing.withsphinx as pws
-        app = util.mock.MagicMock()
-        pws.setup(app)
-        self.assertTrue(mock_required_extensions.called,
-                        self.message.format(times=mock_required_extensions.call_count))
-        mock_required_extensions.assert_called_with(app)
+        self.assertTrue(callable(pws.setup))
+        app = types.ModuleType('a_fatacy_module')
+        self.assertEqual(pws.setup(app), None)
+        self.assertFalse(mock_backports_sphinx15.called,
+                         self.message.format(times=mock_backports_sphinx15.call_count))
+        self.assertFalse(mock_required_extensions.called,
+                         self.message.format(times=mock_required_extensions.call_count))
 
 
 if __name__ == "__main__":
